requirements.txt
# chainlit
llama-index
llama-index-llms-ollama
llama-index-llms-huggingface
llama-index-llms-huggingface-api
llama-index-llms-anthropic
llama_agents
pandas
transformers
accelerate
torch
pyyaml
streamlit